requirements.txt
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.3
django-classy-tags==0.8.0
django-pylibmc==0.6.1
django-pylibmc-sasl==0.2.4
django-taggit==0.22.0
django-taggit-templatetags==0.2.5
django-taggit-templatetags2==1.6.1
django-templatetag-sugar==1.0
django-toolbelt==0.0.1
docutils==0.13.1
draceditor==1.1.6
feedparser==5.2.1
gunicorn==19.6.0
Markdown==2.6.8
markdown2==2.3.3
olefile==0.44
Pillow==4.0.0
psycopg2==2.6.2
pycrypto==2.6.1
Pygments==2.2.0
pygments-style-solarized==0.1.1
pylibmc==1.5.1
python-memcached==1.58
pytz==2016.10
requests==2.13.0
simplegeneric==0.8.1
six==1.10.0
sqlparse==0.2.2
static3==0.7.0
